Temp Login Bypass | Skip Language File | Load Base Attempt
This commit includes directly attempting to create a new user object and bypass the login process. We also skip the network request made to the old servers for the lang.txt file in KEYS.as - it can be assumed that the server returned the appropriate language file depending on the user's nationality and then used those Strings throughout the game - this was removed for simplicity. This commit also included the first look at modelling out all the hardedcoded values that handleBaseLoadSuccessful() requires to function.

NOTE: This commit runs with exceptions