Skipped hashed value check | Attempts to load base | Error message outputs
In this commit, we skipped the hash comparison checks made by the URLLoaderApi class and instead directly trigger the onComplete call - which in turn, allows BASE.as to attempt to load the user's yard via handleBaseLoadSuccessful(). This commit also included modifying the ERRORMESSAGE.as class to output valid error messages to the UI to help in debugging.

NOTE: This commit has undefined errors when running the game.