Server routes now return status 200 to game & URLLoaderApi clear-up
This commit includes a small fix to the server routes by adding a '/' to the end of each route in order to return a status code of 200 and not a 304, which was the original issue. This commit also takes a look at matching the method type that the game expects from /base/load which is POST, to retrieve the data back through GET. Finally, URLLoaderApi class was modified to give the decompiled variable names some meaning and context on what exactly they are.