SingletonLock not found fix in MapRoomManager.as
This commit fixes an issue were the MapRoomManager.as file was not editable in the decompiler as once recompiled, the game cannot find the SingletonLock class at location GAME() according to the ActionScript exception. Classes that have the SingletonLock class at the bottom of their files seem to struggle with this issue and the workaround does not always seem to be the same for all, in this case moving the SingletonLock class to its own file fixed the issue.