Debugger improvements & added new properties to base/load
This commit fixes and improves the debugger for logs to the server. Previously, the request to the server for any given debug log was cut-off by the game erroring out, meaning the logs where convoluted and did not accurately display issues. The new improvement adds all logs to an array and sends them in a single request to the server - avoiding the issue. This commit also sees new properties added to base/load which are required for the game to start. Other minor changes can be seen throughout.

NOTE: This commit contains exceptions in-game