Player base loading successfully
This commit is the first commit which sees the game getting to the point were it loads the player's base successfully for the first time and begins the tutorial phase of the game - this was done by adding the final few properties that handleBaseLoadSuccessful() expects from the server. We can also see the addition of new routes on the server for saving the player's base and setting the map room version.