Loading Map Room 3 & dynamic assets fetched from server at runtime
This commit includes the following changes: Skipped the migration process of prompting the user to 'upgrade' to map room 3 and instead set the current map room version to 3 by default. Created a public route on the server for storing all assets that are dynamically loaded at runtime by the client - this feature DOES NOT work as of this commit as the assets don't load, however, the functionality of requesting those assets from localhost is working and can be monitored via the logs.