All relative paths now append directly to localhost
This commit involved moving all relative paths to insteadd directly append onto localhost. Before this, the relative paths were being appended to file paths in the user's local filesystem; this is due to the fact that the game is being hosted from the SWF file itself and therefore caused it to look at the local filesystem.