Added '/worldmapv3/initworldmap' route & maproom flag
This commit enables maproom 3 by adding a flag on the server which sets maproom2 to equal 1, allowing the maproom to proceed. A route was also added to the server to allow initialization of the worldmap, however, as of this commit, it now throws a new error when accessing the maproom - please see the issues tracker.