Setting coordinates X, Y to 500 x 500 on worldmap v3
This commit adds hardcoded values of 500 by 500 for both the x and y coordinates on the worldmap to load the base onto. This is temporary and will be changed to randomise the value within a range, eventually. Current issue is that the worldmap loads forever, an observation can be seen at _showMapWaiting being set to 1 - check what it does