Important MapRoom3 changes (view commit)
This important commit is the first look at understanding the code associated with the worldmapv3 and identifying the issues. There are some important comments which explain how the maproom gets loaded, renaming of variables for better code clarity, along with some additonal data structures on the server. Our current issue lies with the assets, as it attempts to call MapRoom3WindowHUD() which expects many of the missing assets - commenting out the HUD makes the issue worse.