Fixed 404 'api/bm/getnewmap' route typo on client | Map grid & cells generation server-sided
In this commit a 404 to the server endpoint: 'api/bm/getnewmap' was spotted on the client due to a typo of the route name, this fix led to some new issues that the route needed to handle such as generating each 'cell' or 'point' on the new world map based on its' dimensions, which are set to 500 x 500 by default. The array of cells and dimensions are sent back to the client in the response.