Processing additional data for baseLoad | error handlinng on parsed data
This commit includes the majority of additional data that the client expects from the baseLoad endpoint. Reference for this data was taken from old JSON files of player bases that where downloaded. Error handling was also implemented for when the client attempts a save - this commit ships with error logs, however, the client continues without visible issues.