Refactored KEYS.as to use json language file over .txt implementation | Added en.v8 language
This commit sees the addition of the English language file used for text content on the client - KEYS.as was refactored to accept and process the file as json rather than a text file (.txt). The old implementation was commented out for a more modern and suitable approach. While also seeing some tweaks made to BASE.as and ERRORMESSAGE.as to accept these new changes.