'basesaveid' added to cookie | updateSaved route returning save data | BASE.as clean up
Within this commit we see the use of cookies to persist the basesaveid across routes - updatesaved route now returns the correct data from the database to keep the client in-sync through the use of polling. BASE.as was also cleaned up to improve readability of variable names, specifically in the 'SaveB()', 'handleLoadSuccessful()' and 'Page()' functions.