Creation of Inferno Bases
After defeating Descent, we create a new "inferno"-type save, which is stored to the saves table.

Default saves now generate a baseid_inferno which is assigned to the inferno base after one is obtained.

Loot gained from the descent is stored in the user's main-save "iresources" field, and the resources are linked between the main yard's "iresources" and the inferno yard's "resources"