Chaos Factory/Lab fixes
The initial bug with loading demonstration videos was caused by the "videoURL" to point to an "/assets/video" directory with no URL. This lead to a Security Error, as flash tries to access a local URL instead of a networked one.

This changes the client to search for the demo vids from a URL.

Reproduction demo vids are also included here.